# docs/openapi/resource-center.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new-api Resource Center API",
    "version": "2026-07-17",
    "summary": "Assets, asynchronous image tasks, uploads, and Webhooks.",
    "description": "Public Resource Center API. Async image tasks and uploads use a normal new-api token. Assets use an ak_ resource key. Account-level task Webhooks are configured in the Resource Center console."
  },
...
  "servers": [
    {
      "url": "{base_url}",
      "variables": {
        "base_url": {
          "default": "https://new-api.example.com",
          "description": "Your new-api origin."
        }
      }
    }
  ],
...
  "paths": {
...
    "/v1/assets/{asset_id}": {
      "get": {
        "tags": [
          "Assets"
        ],
        "operationId": "getAsset",
        "summary": "Get an asset",
        "security": [
          {
            "AssetKeyAuth": []
          }
        ],
        "x-required-scope": "assets:read",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "description": "Asset ID such as asset_xxx.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/v1/assets/export": {
      "get": {
        "tags": [
          "Assets"
        ],
        "operationId": "exportAssets",
        "summary": "Export asset URLs as CSV",
        "description": "Exports at most 10,000 matching assets.",
        "security": [
          {
            "AssetKeyAuth": []
          }
        ],
        "x-required-scope": "assets:read",
        "parameters": [
          {
            "name": "asset_type",
            "in": "query",
            "required": false,
            "description": "Filter by asset type.",
            "schema": {
              "$ref": "#/components/schemas/AssetType"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Public APIs return available assets.",
            "schema": {
              "type": "string",
              "enum": [
                "available"
              ]
            }
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "description": "Filter by exact task ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "description": "Filter by exact model.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "description": "Filter by internal task platform name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "description": "Filter by task action.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": false,
            "description": "Filter by channel ID.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "Search asset ID, task ID, filename, or URL.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_timestamp",
            "in": "query",
            "required": false,
            "description": "Minimum creation time in Unix seconds.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "required": false,
            "description": "Maximum creation time in Unix seconds.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV with asset_id, task_id, asset_type, url, filename, model, platform, action, and created_at.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "new-api token",
        "description": "A normal new-api token authorized for the requested image model."
      },
      "AssetKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_*",
        "description": "A Resource Center key. Legacy keys have assets:read only."
      },
      "WebhookBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wk-...",
        "description": "new-api generates this Key for the account-level task Webhook. The account owner can reveal, copy, or regenerate it in the Resource Center."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "code",
          "message"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "invalid_request_error"
          },
          "code": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "AssetType": {
        "type": "string",
        "enum": [
          "image",
          "video",
          "audio",
          "file"
        ]
      },
      "Asset": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "object",
          "id",
          "task_id",
          "index",
          "type",
          "url",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "object": {
            "const": "asset"
          },
          "id": {
            "type": "string",
            "examples": [
              "asset_xxx"
            ]
          },
          "task_id": {
            "type": "string",
            "examples": [
              "task_xxx"
            ]
          },
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "thumbnail_url": {
            "type": "string",
            "format": "uri"
          },
          "mime_type": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "minimum": 0
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "model": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "status": {
            "const": "available"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credential.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Credential, scope, IP, expiry, or user policy denied access.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource was not found for the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}